+++ crates/nostr-sdk/examples/lmdb.rs
@@ -2,7 +2,7 @@
 // Copyright (c) 2023-2025 Rust Nostr Developers
 // Distributed under the MIT software license
 
-use nostr_lmdb::NostrLMDB;
+use nostr_lmdb::NostrLmdb;
 use nostr_sdk::prelude::*;
 
 #[tokio::main]
@@ -11,7 +11,7 @@ async fn main() -> Result<()> {
 
     let keys = Keys::parse("nsec1ufnus6pju578ste3v90xd5m2decpuzpql2295m3sknqcjzyys9ls0qlc85")?;
 
-    let database = NostrLMDB::open("./db/nostr-lmdb").await?;
+    let database = NostrLmdb::open("./db/nostr-lmdb").await?;
     let client: Client = ClientBuilder::default()
         .signer(keys.clone())
         .database(database)