+++ database/nostr-lmdb/CHANGELOG.md
@@ -31,6 +31,7 @@
 
 ### Changed
 
+- Rename `NostrLMDB` to `NostrLmdb`
 - Spawn blocking task for read operations (https://github.com/rust-nostr/nostr/pull/1141)
 
 ## v0.44.0 - 2025/11/06